Move quick links settings component inside quick-links feature
